Create parent class Entity
